+++ examples/all_usage_example.py
@@ -1,8 +1,9 @@
+from time import time
 from imdbinfo.services import search_title, get_movie, get_name
 
 import logging
 logging.basicConfig(level=logging.WARNING)
-
+# list_q = "little house|matrix|walking dead|horizon".split("|")
 list_q = "France to find his estate confiscated by governor Narbonne, for back taxes, and resold to Katrina, a Dutch Countess. Katrina offers to return Pierre's property if he will help her get possession".split(" ")
 #Example 1: Search for a movie by title
 
@@ -13,13 +14,20 @@ for word in list_q:
         for movie in results.titles:
             print(f"{movie.title} ({movie.year}) - {movie.imdb_id}")
             movie = get_movie(movie.imdb_id)
-            print(f"  Valutazione: {movie.rating}")
+            print(f"  Rating: {movie.rating}")
+            print(f"  Kind: {movie.kind}")
+            if movie.is_series():
+                print(f"  Series Info: {movie.info_series or 'N/A'}")
+            if movie.is_episode():
+                print(f"  Episode Info: {movie.info_episode or 'N/A'}")
             for director in movie.directors:
                 print(f"  Director: {director.name}")
-            for c in movie.categories['cast'][:3]:  # Limit to first 3 cast members for brevity
+            for c in movie.categories['cast'][:1]:  # Limit to first 3 cast members for brevity
                 #print(c)
+                t0 = time()
                 person = get_name(c.imdb_id)
-                print(f"  Cast: {person.name} - {' ° '.join(person.primary_profession[:2])} - {person.url}")
+                t1 = time()
+                print(f"  Cast: \n\t{person.name} ({', '.join(c.characters)}) - {' ° '.join(person.primary_profession[:2])} - {person.url} ")
             # DONE
             print("----------------------------------------------")