+++ imdbinfo/models.py
@@ -1,8 +1,57 @@
+import datetime
 from typing import Optional, List, Dict, Tuple, Union
 from pydantic import BaseModel, field_validator
 import logging
 
+from imdbinfo.transformers import _release_date
+
+EPISODE_IDENTIFIERS = ("tvEpisode", "podcastEpisode")
+
+SERIES_IDENTIFIERS = ("tvSeries", "tvMiniSeries", "podcastSeries")
+
 logger = logging.getLogger(__name__)
+
+class SeriesMixin:
+    def is_series(self) -> bool:
+        """
+        Check if this movie title is a series, the main title of a series.
+        If True, it means that this is a series, not a movie, not an episode, but the main reference for the series itself, and series details can be found in the self.info_series property.
+        """
+        return getattr(self, "kind", None) in SERIES_IDENTIFIERS
+
+    def is_episode(self) -> bool:
+        """
+        Check if this movie title is an episode of a series.
+        If True, means that this is the episode of a series and episode details can be found in the self.info_episode property
+        """
+        return getattr(self, "kind", None) in EPISODE_IDENTIFIERS
+
+class InfoSeries(BaseModel):
+    display_years : List[str] = []  # e.g. ['2013', '2014', '2015']
+    display_seasons : List[str] = []  # e.g. ['1', '2', '3']
+
+    @field_validator('display_years', mode='before')
+    def filter_years(cls, value):
+        if value is None:
+            return []
+        return [str(y) for y in value if isinstance(y, str) and len(y) == 4 and y.isdigit()]
+
+    def __str__(self):
+        return f"Years: {self.display_years[-1] if self.display_years else ''}-{self.display_years[0] if self.display_years else ''}, Seasons: {len(self.display_seasons)}"
+
+
+class InfoEpisode(BaseModel):
+    season_n: Optional[int] = None
+    episode_n: Optional[int] = None
+    series_imdbId: Optional[str] = None
+    series_title: Optional[str] = None
+    series_title_localized: Optional[str] = None
+
+    def __str__(self):
+        # print in S01E01 format
+        season_str = f"S{self.season_n:02d}" if self.season_n is not None else "S??"
+        episode_str = f"E{self.episode_n:02d}" if self.episode_n is not None else "E??"
+        return f"{self.series_title} - {season_str}{episode_str} ({self.series_imdbId})"
 
 class Person(BaseModel):
     """ person model for directors, cast and search results.
@@ -90,7 +139,7 @@ class CastMember(Person):
         return f"{self.name} ({', '.join(self.characters)})"
 
 
-class MovieDetail(BaseModel):
+class MovieDetail(SeriesMixin, BaseModel):
     """MovieDetail model for detailed information about a movie.
     This model contains all the information about a movie such as title, id, imdb_id, imdbId, url, cover_url, plot, release_date, languages, certificates, directors, stars,
     year, duration, country_codes, rating, metacritic_rating, votes, trailers, genres, interests, worldwide_gross, production_budget, storyline_keywords,
@@ -115,6 +164,7 @@ class MovieDetail(BaseModel):
     directors: List[Person] = []
     stars: List[Person] = []
     year: Optional[int] = None
+    year_end: Optional[int] = None
     duration: Optional[int] = None
     country_codes: List[str] = []
     rating: Optional[float] = None
@@ -140,16 +190,25 @@ class MovieDetail(BaseModel):
     production: List[str] = []
     categories: Dict[str, List[Union[Person, CastMember]]] = {}
 
+
     @field_validator('languages', 'country_codes','genres', mode='before')
     def none_is_list(cls, value):
         if value is None:
             return []
         return value
 
+    def __str__(self):
+        return f"{self.title} ({self.year}) - {self.imdbId} ({self.kind})"
 
-class MovieInfo(BaseModel):
+class TvSeriesDetail(MovieDetail):
+    info_series: Optional[InfoSeries] = None  # e.g. SeriesInfo(display_years=['2013', '2014', '2015'], display_seasons=['1', '2', '3'])
+
+class TvEpisodeDetail(MovieDetail):
+    info_episode: Optional[InfoEpisode] = None  # e.g. SeriesInfo(display_year
+
+class MovieBriefInfo(SeriesMixin, BaseModel):
     """
-    MovieInfo model for search results and cast members.
+    MovieBriefInfo model for search results and cast members.
     This model is used to represent a movie in search results and cast members.
     It contains basic information about a movie such as title, id, imdb_id, imdbId, url, cover_url, year and kind.
     It can be used to represent a movie in search results or as part of a cast member's credits.
@@ -162,7 +221,7 @@ class MovieInfo(BaseModel):
     cover_url: Optional[str] = None
     url: Optional[str] = None
     year: Optional[int] = None # TODO series will have year as string 'from-to'. For now only movies are supported
-    kind: Optional[str] = None # e.g. 'movie', 'series', 'episode', 'video game', etc.
+    kind: Optional[str] = None # e.g. 'movie', 'tvSeries', 'tvSeriesEpisode' ...
 
 
     @classmethod
@@ -182,8 +241,8 @@ class MovieInfo(BaseModel):
         )
 
     @classmethod
-    def from_cast(self, data: dict):
-        return self(
+    def from_cast(cls, data: dict):
+        return cls(
             id=str(data['id'].replace('tt', '')),
             imdb_id=str(data['id'].replace('tt', '')),
             imdbId=data['id'],
@@ -199,9 +258,9 @@ class SearchResult(BaseModel):
     SearchResult model for search results.
     This model contains the results of a search query, including a list of titles and names.
     It is used to represent the results of a search query for movies and people.
-    It includes a list of MovieInfo objects for titles and a list of Person objects for names.
+    It includes a list of MovieBriefInfo objects for titles and a list of Person objects for names.
     """
-    titles: List[MovieInfo] = []
+    titles: List[MovieBriefInfo] = []
     names: List[Person] = []
 
 
@@ -228,10 +287,124 @@ class PersonDetail(BaseModel):
     death_place: Optional[str] = None
     death_reason: Optional[str] = None
     jobs: List[str] = []
-    credits: Dict[str, List[MovieInfo]] = {}
-    unreleased_credits: Dict[str, List[MovieInfo]] = {}
+    credits: Dict[str, List[MovieBriefInfo]] = {}
+    unreleased_credits: Dict[str, List[MovieBriefInfo]] = {}
 
 
     def __str__(self):
         return f"{self.name} ({', '.join(self.knownfor)})"
 
+class SeasonEpisode(BaseModel):
+    id: str  # id without 'tt' prefix, e.g. '1234567'
+    imdbId: str
+    imdb_id: str
+    title: str
+    season: int
+    episode: int
+    plot: str
+    image_url: Optional[str] = None
+    rating: Optional[float] = None
+    votes: Optional[int] = None
+    year: Optional[int] = None
+    release_date: Optional[str] = None
+    kind: Optional[str] = None
+
+    @classmethod
+    def from_episode_data(cls, data: dict) -> 'SeasonEpisode':
+        """
+        Create a SeasonEpisode instance from episode data dictionary.
+        """
+        return cls(
+            id=data['id'].replace('tt', ''),
+            imdbId=data['id'],
+            imdb_id=data['id'].replace('tt', ''),
+            title=data['titleText'],
+            season=data['season'],
+            episode=data['episode'],
+            plot=data.get('plot',''),
+            image_url=data.get('image', {}).get('url', None),
+            rating=data.get('aggregateRating', None),
+            votes=data.get('voteCount', None),
+            year=data.get('releaseYear', None),
+            release_date=_release_date(data['releaseDate']),
+            kind=data.get('type'),
+
+        )
+
+
+    def __str__(self):
+        return f"{self.title} (S{self.season:02d}E{self.episode:02d}) - {self.imdbId} ({self.year or 'N/A'}) - {self.kind or 'N/A'}"
+
+
+class BulkedEpisode(BaseModel):
+    id: str  # id without 'tt' prefix, e.g. '1234567'
+    imdbId: str
+    imdb_id: str
+    title: str
+    plot: str
+    image_url: Optional[str] = None
+    rating: Optional[float] = None
+    votes: Optional[int] = None
+    year: Optional[int] = None
+    release_date: Optional[str] = None
+    kind: Optional[str] = None
+    genres: Optional[List[str]] = None
+    duration: Optional[int] = None  # Duration in seconds
+
+    @classmethod
+    def from_bulked_episode_data(cls, data: dict) -> 'BulkedEpisode':
+        """
+        Create an EpisodeData instance from bulked episode data dictionary.
+        This is used when fetching episodes in bulk from a series.
+        """
+        return cls(
+            id=data['titleId'].replace('tt', ''),
+            imdbId=data['titleId'],
+            imdb_id=data['titleId'].replace('tt', ''),
+            title=data['titleText'],
+            genres= data.get('genres') or [],
+            plot=data.get('plot',''),
+            image_url=data.get('primaryImage', {}).get('url', None),
+            rating=data.get('ratingSummary', {}).get('aggregateRating', None),
+            votes=data.get('ratingSummary', {}).get('voteCount', None),
+            year=data.get('releaseYear', None),
+            release_date=_release_date(data['releaseDate']),
+            kind=data.get('titleType',{}).get('id', None),
+            duration=data.get('runtime'),
+
+        )
+
+    def __str__(self):
+        return f"{self.title} ({self.release_date or 'N/A'}) - {self.imdbId} ({self.kind or 'N/A'})"
+
+
+
+class SeasonEpisodesList(BaseModel):
+    """
+    EpisodesList model for a list of episodes.
+    This model contains a list of EpisodeInfo objects representing the episodes of a series.
+    It can be used to represent the episodes of a series in a specific season.
+    """
+    top_rating_episode :  Optional[float] = None
+    total_series_episodes  : Optional[int] = None  # Total number of episodes in the series
+    total_series_seasons : Optional[int] = None  # Total number of seasons in the series
+    top_ten_episodes : Optional[List[dict]]  = None # List of top ten episodes based on rating
+    episodes: List[SeasonEpisode] = []
+
+    @property
+    def count(self)-> int:
+        """
+        Count the number of episodes in the list.
+        Returns:
+            int: The number of episodes in the list.
+        """
+        return len(self.episodes)
+
+    def __len__(self):
+        return len(self.episodes)
+
+    def __getitem__(self, idx):
+        return self.episodes[idx]
+
+    def __str__(self):
+        return f"Total Episodes: {len(self.episodes)}"